# feat(rslint_parser): Avoid cloning parser state