… crates/rslint_parser/src/syntax/expr.rs
… crates/rslint_parser/src/syntax/function.rs
… crates/rslint_parser/src/syntax/module.rs
… crates/rslint_parser/src/syntax/object.rs
… crates/rslint_parser/src/syntax/program.rs
… crates/rslint_parser/src/syntax/stmt.rs
… crates/rslint_parser/src/syntax/typescript.rs