## Summary

The parser has a neat `with_state` function that takes a snapshot of the parser state and restores the state to its previous state when it goes out of scope. However, the downside is that it copies the state at least twice, once inside the `with_state` function and once in the call-site, when constructing the new state which results in a serious performance penalty. 

This PR introduces a new `ChangeParserState` trait that allows making specific changes to the `ParserState` that can later be undone, only keeping a copy of the data they changed (and need to be restored later). Ideally, these changes would have semantic names. However, this PR re-uses the name of the fields. 

I hope to be able to reduce a lot of the boilerplate by introducing generic fields, for example if we're inside a function, constructor, async, etc... that then can all be changed with a single `ChangeParserState` implementation.




## Performance

The parse time (excluding tree sink and lexing) or across all libs improves by 25-33%, in some cases even close to 50%

* jquery.min.js: 3.244868ms -> 2.395019ms
* vue.global.prod.js: 6.825166ms -> 3.757078ms
* react.production.min.js: 246.53µs -> 191.09µs
* react-dom.production.min.js: 4.974296ms -> 3.006538ms
* compiler.js: 12.915072ms -> 9.110865ms
* dojo.js: 784.669µs -> 617.38µs
* d3.min.js: 11.771383ms -> 8.440315ms
* pixi.min.js: 12.437823ms -> 10.148374ms
* three.min.js:  19.487128ms -> 11.145823ms
* tex-chtml-full.js: 29.200492ms  -> 21.564177ms

## Test Plan

Verified that coverage doesn't change.