@ematipico 2022-01-11
> However, I plan to group many of the flags fields in a bit-flag and then introduce a with_flags function that implements the old behaviour but in a way so that it only snapshots the flags which is the same as storing an integer.

Would you mind create an issue and explain there, in more detail, what's the approach to plan to design?