@Boshen 2022-01-11
From my understanding of Rust, I think the pattern we need to explore here is the concept of Interior Mutability, i.e.`std::cell::Cell`, and use `std::cell::swap/replace`. We can then remove &mut on parser to make it clear what can be mutated. 

```rust
pub struct ParserState {
	pub strict: Rc<Option<StrictMode>>,
}
```

```rust
	let mut last_strict = Some(StrictMode::Class(p.cur_tok().range()));
	std::mem::swap(&mut p.state.strict, &mut last_strict);

becomes

	let last_strict = Cell::new(Some(StrictMode::Class(p.cur_tok().range())));
	p.state.strict.swap(last_strict);
```