@leops 2022-01-11
I think there's a workable design in-between both the previous drop guard based design and the new save and restore behavior:

```rust
trait ParserModifier {
    fn enter(state: &mut ParserState) -> Self;
    fn exit(self, state: &mut ParserState);
}

struct ParserStateGuard<'state, T: ParserModifier> {
    modifier: T,
    state: &mut ParserState,
}

impl<'state, T: ParserModifier> ParserStateGuard<'state, T> {
    fn enter(state: &mut ParserState) -> Self {
        Self { modifier: T::enter(state), state }
    }
}

impl<'state, T: ParserModifier> Drop for ParserStateGuard<T> {
    fn drop(&mut self) {
        T::exit(self.modifier, self.state);
    }
}

struct InGenerator(bool);

impl ParserModifier for InGenerator {
    fn enter(state: &mut ParserState) -> Self {
        Self(replace(&mut state.in_generator, true))
    }

    fn exit(self, state: &mut ParserState) {
        state.in_generator = self.0;
    }
}
```

This is an approximate example, I haven't included the implementation of `Deref` for `ParserStateGuard`. Also it would be possible to have generic implementations of `ParserModifier` over tuples of modifiers to facilitate merging multiple changes at once