@MichaReiser 2022-01-11
> > However, I plan to group many of the flags fields in a bit-flag and then introduce a with_flags function that implements the old behaviour but in a way so that it only snapshots the flags which is the same as storing an integer.
> 
> Would you mind create an issue and explain there, in more detail, what's the approach to plan to design?

I don't know yet in detail but it's the next I plan to tackle. I was thinking of grouping some of the flags like `in_async`, `in_generator` together as well as `in_function` and `in_constructor` but I still need to work out the details because JS allows for some weird constructs. 

> I think there's a workable design in-between both the previous drop guard based design and the new save and restore behavior:

That's actually something I wanted to look in next once I figure out how we want to restructure the parser state. But let me give it a try and fine tune it from their. One difference is that this design doesn't allow to restore the state on drop but I think that's fine.