@xunilrj 2022-01-11
> I think there's a workable design in-between both the previous drop guard based design and the new save and restore behavior:

I liked your approach. Remind me of "mutable lens" with undo.

The only crate that tries to do "mutable lens", as far as I know, is https://github.com/nsmryan/myopic. But there are not a lot of examples and its use seems harder than necessary. All other "lens" (linked at myopic readme) focus on "immutability with clone", what we trying to avoid here.

From the undo/redo crates, I think this one is closer to what we need. Apart from the "Point!" macro in the example is very close to what we want.
https://github.com/M-Adoo/rundo/blob/master/docs/quickstart.md