@MichaReiser 2022-01-11
I implemented @leops proposal but had to make two changes:

* The `apply` method making the change to the `ParserState` must be an instance method so that it can be parametrized. For example, setting the `strict` field requires knowing the range of the `use strict` directive. 
* A consequence of the above is that there are now two structs for each change. The struct implementing the change logic and a new-type wrapper for storing the previous state. 

@xunilrj this crate looks interesting but also feels a bit magic. There are also fields that are tracked and untracked at the same time. For example, the `name_map` field must not be tracked when items are added or removed to it but there are places where we replace the whole map, which should be tracked. I further hope that we can come up with more semantic names for the changes, so that these become more than simple "mutation calls".