@MichaReiser 2022-01-11
> ### Test262 comparison coverage results on ubuntu-latest
> Test result 	`main` count 	This PR count 	Difference
> Total 	44951 	44951 	0
> Passed 	42923 	42923 	0
> Failed 	2027 	2027 	0
> Panics 	1 	1 	0
> Coverage 	95.49% 	95.49% 	0.00%

Thanks @Boshen for fixing the ubuntu test reports!