@xunilrj 2022-01-11
I did a quick test here and it seems to be working. What do you think?
It alleviates the boilerplate needed, and it makes more obvious that all behave the same.

```rust
macro_rules! gen_parser_mutate {
	($name:ident { $field:ident: $type:path} ) => {
		/// Changes the [ParserState] `in_generator` field
		#[derive(Debug)]
		pub struct $name(pub (crate) $type);

		impl ChangeParserState for $name {
			type Snapshot = $type;

			fn apply(self, state: &mut ParserState) -> Self::Snapshot {
				std::mem::replace(&mut state.$field, self.0)
			}

			fn restore(state: &mut ParserState, value: Self::Snapshot) {
				state.$field = value
			}
		}
	};
}

gen_parser_mutate!{InGenerator{ in_generator: bool}}
gen_parser_mutate!{InFunction{ in_function: bool}}
gen_parser_mutate!{InAsync{ in_async: bool}}
```

It is only a little bit different from yours.
example:


```rust
let p = &mut *p.with_state(
    InFunction(true)
	.and(InGenerator(in_generator))
	.and(InAsync(in_async))
);
```