@NicholasLYang 2022-01-11
Perhaps this is a dumb question, but is there a way to have the API use a closure? I've had functions that allow you to run code in a specific state, then restore once the closure finishes