@MichaReiser 2022-01-11
> Perhaps this is a dumb question, but is there a way to have the API use a closure? I've had functions that allow you to run code in a specific state, then restore once the closure finishes.
> 
> The reason I ask is that having the change occur on drop would likely be a little confusing. If we can have an API like:
> 
> ```
> run_in_state!(in_generator: true, |p| {  ... })
> ```

That would work as well. I guess, the main question is how you type the ‚in_generator‘. We could use the same approach as introduced in this diff so that ‚run_with_state‘ is just a different api to ‚with_state‘. 

Another API would be a parserstate with all fields being options but this then requires finding the initialized fields, which may be a bit cumbersome.