@NicholasLYang 2022-01-12
This is where macros could work quite well no?  Like we can have:

```
macro_rules! run_in_state
(($field:ident: $value:expr)+, $func:expr) => {
  // swap fields
  let fn = $func;
  fn();
  // swap back fields
}
```